[.files/lib/] tellfiles/ deleted.
Historically, the so-called 'tell-files' were empirically intended to serve two purposes. At first, moments before their creation, it was only a single .DS_Store file made by Finder in a freshly-set open window of the user home of a factory-installed stock OS X Mavericks, which was saved to later be copied in locations that would have to be set. As much as that solution was disliked, it was the only immediately available option. The same day, it had been needed to create multiple .nomedia files in the user home of an Android Lollipop. At that time, it was planned to write a .android setup script, and the idea of writing files such as .hushlogin or .nomedia directly via setup scripts was already due to be implemented. By then many times already had arisen the problem of complex software setup automatization; such software saves its configuration in quirky, often proprietary files, way more than just filenames to detect. That kind of files were naively coined 'tell-files', for files that tell something to select software and not just say things to anywho, and so was it decided to build a library of ready-to-deploy preset tell-files... To forthwith abandon it. Setting up by importing presets is a hard way to configure anything. Like any hard way, it tends to become unreliable in the long term. Unreadable software-generated files are hardly rewritable. Also updated proprietary software may update only its files detected in the update nay not at all. Quality design seldom being the developers' concern, not even speaking of reliability, many programs crash when reading corrupt or incompatible special files. It is so full of fail that even hacked or reverse-engineered configuration is a softer approach, potentially more reliable if the code writer is careful. To avoid such problems, each tell-file shall be recreated from scratch with each version of its genetic software, stored apart and carefully handled. Even if there was a script which would select the right one, another problem would soon arise: the space this growing archive takes. Improved junk is still junk. For all these reasons, the tell-files library project had been immediately canceled. This update only cleans up the forgotten leftovers.

[Comme dit Papa, 'Avec un âne, tu fais pas un cheval de course'. Ça part à la casse. Hop.]